Move smithay code from `gtk-capture` example into `cosmic-client-toolkit`
