+++ client-toolkit/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 cosmic-protocols = { path = "../" }
 sctk = { package = "smithay-client-toolkit", git = "https://github.com/Smithay/client-toolkit", rev = "8b47ecbc" }
 wayland-client = { version = "0.30.0-beta.10" }
+smithay = { git = "https://github.com/Smithay/smithay", rev = "625cbca5", default-features = false, features = ["backend_egl", "renderer_gl", "renderer_multi", "slog-stdlog"], optional = true }
 
 [dev-dependencies]
 async-channel = "1.7.1"
@@ -15,5 +16,11 @@ env_logger = "0.9.1"
 gdk-pixbuf = "0.15.11"
 gtk4 = { version = "0.4.8" }
 gdk4-wayland = { version = "0.4.8", features = ["wayland_crate", "v4_4", "egl"] }
-smithay = { git = "https://github.com/Smithay/smithay", rev = "625cbca5", default-features = false, features = ["backend_egl", "renderer_gl", "renderer_multi", "slog-stdlog"] }
 wayland-backend = { version = "0.1.0-beta.10", features = ["client_system"] }
+
+[features]
+default = ["smithay"]
+
+[[example]]
+name = "gtk-capture"
+required-features = ["smithay"]
